+++ basictracer/propagation.py
@@ -1,36 +1,34 @@
 from __future__ import absolute_import
 
-import base64
-import time
 import struct
-from opentracing import InvalidCarrierException, TraceCorruptedException
-from .context import Context
-from .span import BasicSpan
+from opentracing import InvalidCarrierException, SpanContextCorruptedException
+from .context import SpanContext
 from .wire_pb2 import TracerState
-from .util import generate_id
 
-_proto_size_bytes = 4 # bytes
+_proto_size_bytes = 4  # bytes
+
 
 class BinaryPropagator(object):
+
     def __init__(self, tracer):
         self.tracer = tracer
 
-    def inject(self, span, carrier):
+    def inject(self, span_context, carrier):
         if type(carrier) is not bytearray:
             raise InvalidCarrierException()
         state = TracerState()
-        state.trace_id = span.context.trace_id
-        state.span_id = span.context.span_id
-        state.sampled = span.context.sampled
-        if span.context.baggage is not None:
-            for key in span.context.baggage:
-                state.baggage_items[key] = span.context.baggage[key]
+        state.trace_id = span_context.trace_id
+        state.span_id = span_context.span_id
+        state.sampled = span_context.sampled
+        if span_context.baggage is not None:
+            for key in span_context.baggage:
+                state.baggage_items[key] = span_context.baggage[key]
 
         # The binary format is {uint32}{protobuf} using big-endian for the uint
-        carrier.extend(struct.pack(">I", state.ByteSize()))
+        carrier.extend(struct.pack('>I', state.ByteSize()))
         carrier.extend(state.SerializeToString())
 
-    def join(self, operation_name, carrier):
+    def extract(self, carrier):
         if type(carrier) is not bytearray:
             raise InvalidCarrierException()
         state = TracerState()
@@ -39,14 +37,11 @@ class BinaryPropagator(object):
         for k in state.baggage_items:
             baggage[k] = state.baggage_items[k]
 
-        return BasicSpan(self.tracer,
-                operation_name=operation_name,
-                start_time=time.time(),
-                context=Context(span_id=generate_id(),
-                    parent_id=state.span_id,
-                    trace_id=state.trace_id,
-                    baggage=baggage,
-                    sampled=state.sampled))
+        return SpanContext(
+            span_id=state.span_id,
+            trace_id=state.trace_id,
+            baggage=baggage,
+            sampled=state.sampled)
 
 
 prefix_tracer_state = 'ot-tracer-'
@@ -56,19 +51,21 @@ field_name_span_id = prefix_tracer_state + 'spanid'
 field_name_sampled = prefix_tracer_state + 'sampled'
 field_count = 3
 
+
 class TextPropagator(object):
+
     def __init__(self, tracer):
         self.tracer = tracer
 
-    def inject(self, span, carrier):
-        carrier[field_name_trace_id] = '{:x}'.format(span.context.trace_id)
-        carrier[field_name_span_id] = '{:x}'.format(span.context.span_id)
-        carrier[field_name_sampled] = str(span.context.sampled).lower()
-        if span.context.baggage is not None:
-            for k in span.context.baggage:
-                carrier[prefix_baggage+k] = span.context.baggage[k]
+    def inject(self, span_context, carrier):
+        carrier[field_name_trace_id] = '{:x}'.format(span_context.trace_id)
+        carrier[field_name_span_id] = '{:x}'.format(span_context.span_id)
+        carrier[field_name_sampled] = str(span_context.sampled).lower()
+        if span_context.baggage is not None:
+            for k in span_context.baggage:
+                carrier[prefix_baggage+k] = span_context.baggage[k]
 
-    def join(self, operation_name, carrier):
+    def extract(self, carrier):
         count = 0
         span_id, trace_id, sampled = (0, 0, False)
         baggage = {}
@@ -87,19 +84,16 @@ class TextPropagator(object):
                 elif v == str(False).lower():
                     sampled = False
                 else:
-                    raise TraceCorruptedException()
+                    raise SpanContextCorruptedException()
                 count += 1
             elif k.startswith(prefix_baggage):
                 baggage[k[len(prefix_baggage):]] = v
 
         if count != field_count:
-            raise TraceCorruptedException()
+            raise SpanContextCorruptedException()
 
-        return BasicSpan(self.tracer,
-                operation_name=operation_name,
-                start_time=time.time(),
-                context=Context(span_id=generate_id(),
-                    parent_id=span_id,
-                    trace_id=trace_id,
-                    baggage=baggage,
-                    sampled=sampled))
+        return SpanContext(
+            span_id=span_id,
+            trace_id=trace_id,
+            baggage=baggage,
+            sampled=sampled)