+++ basictracer/tracer.py
@@ -2,11 +2,12 @@ import time
 
 from opentracing import Format, Tracer
 from opentracing import UnsupportedFormatException
-from .context import Context
+from .context import SpanContext
 from .propagation import BinaryPropagator, TextPropagator
 from .recorder import SpanRecorder, DefaultSampler
 from .span import BasicSpan
 from .util import generate_id
+
 
 class BasicTracer(Tracer):
 
@@ -17,48 +18,62 @@ class BasicTracer(Tracer):
         self._text_propagator = TextPropagator(self)
         return
 
-    def start_span(self,
+    def start_span(
+            self,
             operation_name=None,
-            parent=None,
+            references=None,
             tags=None,
             start_time=None):
+
         start_time = time.time() if start_time is None else start_time
-        context = Context(span_id=generate_id())
-        sp = BasicSpan(self, operation_name=operation_name, tags=tags,
-                context=context,
-                start_time=start_time)
 
-        if parent is None:
-            sp.context.trace_id = generate_id()
-            sp.context.sampled = self.sampler.sampled(sp.context.trace_id)
+        # See if we have a parent_ctx in `references`
+        parent_ctx = None
+        if references is not None:
+            references = (
+                references if isinstance(references, list)
+                else [references])
+            parent_ctx = references[0].span_context
+
+        # Assemble the child ctx
+        ctx = SpanContext(span_id=generate_id())
+        if parent_ctx is not None:
+            if parent_ctx.baggage is not None:
+                ctx.baggage = parent_ctx.baggage.copy()
+            ctx.trace_id = parent_ctx.trace_id
+            ctx.sampled = parent_ctx.sampled
         else:
-            sp.context.trace_id = parent.context.trace_id
-            sp.context.parent_id = parent.context.span_id
-            sp.context.sampled = parent.context.sampled
-            if parent.context.baggage is not None:
-                sp.context.baggage = parent.context.baggage.copy()
+            ctx.trace_id = generate_id()
+            ctx.sampled = self.sampler.sampled(ctx.trace_id)
 
-        return sp
+        # Tie it all together
+        return BasicSpan(
+            self,
+            operation_name=operation_name,
+            context=ctx,
+            parent_id=(None if parent_ctx is None else parent_ctx.span_id),
+            tags=tags,
+            start_time=start_time)
 
-
-    def inject(self, span, format, carrier):
+    def inject(self, span_context, format, carrier):
         if format == Format.BINARY:
-            self._binary_propagator.inject(span, carrier)
+            self._binary_propagator.inject(span_context, carrier)
         elif format == Format.TEXT_MAP:
-            self._text_propagator.inject(span, carrier)
+            self._text_propagator.inject(span_context, carrier)
         else:
             raise UnsupportedFormatException()
 
-    def join(self, operation_name, format, carrier):
+    def extract(self, format, carrier):
         if format == Format.BINARY:
-            return self._binary_propagator.join(operation_name, carrier)
+            return self._binary_propagator.extract(carrier)
         elif format == Format.TEXT_MAP:
-            return self._text_propagator.join(operation_name, carrier)
+            return self._text_propagator.extract(carrier)
         else:
             raise UnsupportedFormatException()
 
     def record(self, span):
         self.recorder.record_span(span)
+
 
 class NoopRecorder(SpanRecorder):
     def record_span(self, span):