+++ tests/test_propagation.py
@@ -1,41 +1,43 @@
 import pytest
-from opentracing import Format, UnsupportedFormatException
+from opentracing import ChildOf, Format, UnsupportedFormatException
 from basictracer import BasicTracer
+
 
 def test_propagation():
     tracer = BasicTracer()
-    sp = tracer.start_span(operation_name="test")
+    sp = tracer.start_span(operation_name='test')
     sp.context.sampled = False
-    sp.set_baggage_item("foo", "bar")
+    sp.context.set_baggage_item('foo', 'bar')
     opname = 'op'
 
     # Test invalid types
     with pytest.raises(UnsupportedFormatException):
-        tracer.inject(sp, "invalid", {})
+        tracer.inject(sp.context, 'invalid', {})
     with pytest.raises(UnsupportedFormatException):
-        tracer.join("", "invalid", {})
+        tracer.extract('invalid', {})
 
     tests = [(Format.BINARY, bytearray()),
              (Format.TEXT_MAP, {})]
     for format, carrier in tests:
-        tracer.inject(sp, format, carrier)
+        tracer.inject(sp.context, format, carrier)
+        extracted = tracer.extract(format, carrier)
 
-        child = tracer.join(opname, format, carrier)
+        assert extracted.trace_id == sp.context.trace_id
+        assert extracted.span_id == sp.context.span_id
+        assert extracted.sampled == sp.context.sampled
+        assert extracted.baggage == sp.context.baggage
 
-        assert child.context.trace_id == sp.context.trace_id
-        assert child.context.parent_id == sp.context.span_id
-        assert child.context.sampled == sp.context.sampled
-        assert child.context.baggage == sp.context.baggage
 
 def test_start_span():
     """ Test in process child span creation."""
     tracer = BasicTracer()
-    sp = tracer.start_span(operation_name="test")
-    sp.set_baggage_item("foo", "bar")
+    sp = tracer.start_span(operation_name='test')
+    sp.context.set_baggage_item('foo', 'bar')
 
-    child = tracer.start_span(operation_name="child", parent=sp)
+    child = tracer.start_span(
+            operation_name='child', references=ChildOf(sp.context))
 
     assert child.context.trace_id == sp.context.trace_id
-    assert child.context.parent_id == sp.context.span_id
     assert child.context.sampled == sp.context.sampled
     assert child.context.baggage == sp.context.baggage
+    assert child.parent_id == sp.context.span_id